Add Secure Sharing UI + deniable identity store (Phases 5 & 6)
SecureSharing coordinator: owns the wallet-style identity (BIP39 seed) and peer
list, persisted as a single opaque, label-less encrypted blob with a generic
filename (store.dat) — deniable at rest. Drives the relay: register, host the
unlocked vault to approved peers only, browse a peer, and query the audit log.
RelayClient gains a signed audit-query call.

UI (MainWindow.SecureSharing.cs, dialog-driven to avoid XAML-load risk):
create/recover identity (shows the recovery phrase once), show ID + safety
number with copy, add peers, go online/offline (hosts the current unlocked
vault), browse a peer's shared vault and open media (fetched to a temp file via
the existing viewer), and an access-log viewer. Settings gains a "Secure
sharing" section: relay URL (AppState.SecureRelayUrl) + a Manage button.

Relay timestamps switched to integer unix seconds so the Ed25519-signed
message strings match byte-for-byte between C# and Python.

Verified: full service stack passes an integration loopback test against the
live relay (deniable store create/open/wrong-pass + no plaintext labels, peer
approval, online hosting, E2E browse byte-exact, unapproved-peer rejection,
opaque audit with no filename leak). App builds clean and launches without a
XAML-load crash. README updated.

+++ server/relay.py
@@ -82,7 +82,7 @@ def verify_sig(sign_pub_b64: str, message: bytes, sig_b64: str) -> bool:
         return False
 
 
-def fresh(ts: float, skew: int = 300) -> bool:
+def fresh(ts: int, skew: int = 300) -> bool:
     """Reject stale/forward-dated timestamps (replay window)."""
     return abs(time.time() - ts) <= skew
 
@@ -93,7 +93,7 @@ class RegisterReq(BaseModel):
     uuid: str
     sign_pub: str   # base64
     agree_pub: str  # base64
-    ts: float
+    ts: int         # unix seconds (integer — so the signed-message string matches byte-for-byte across languages)
     sig: str        # base64 Ed25519 over "register:{uuid}:{sign_pub}:{agree_pub}:{ts}"
 
 
@@ -129,7 +129,7 @@ def lookup(uuid: str):
 
 class AuditQuery(BaseModel):
     uuid: str
-    ts: float
+    ts: int
     sig: str  # Ed25519 over "audit-query:{uuid}:{ts}"
 
 
@@ -182,7 +182,7 @@ async def connect(ws: WebSocket):
     try:
         auth = json.loads(await ws.receive_text())
         uuid = str(auth.get("uuid", ""))
-        ts = float(auth.get("ts", 0))
+        ts = int(auth.get("ts", 0))
         with closing(db()) as conn:
             peer = conn.execute("SELECT sign_pub FROM peers WHERE uuid=?", (uuid,)).fetchone()
         if not peer or not fresh(ts) or not verify_sig(peer["sign_pub"], f"connect:{uuid}:{ts}".encode(), auth.get("sig", "")):
